add/move: DbReset/Database classes to testsupport package
