+++ src/test/java/com/foodtruck/testsupport/Database.java
@@ -1,4 +1,4 @@
-package com.foodtruck.util;
+package com.foodtruck.testsupport;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
